20151022: Remove the -refactored suffixes; changed default number of threads.
